+++ fidesctl/src/fidesapi/database/crud.py
@@ -2,26 +2,17 @@
 Contains all of the generic CRUD endpoints that can be
 generated programmatically for each resource.
 """
-# pylint: disable=redefined-outer-name,cell-var-from-loop
-
 from typing import Dict, List
 
-from fastapi import APIRouter, status
 from loguru import logger as log
 from sqlalchemy import update as _update, delete as _delete
 from sqlalchemy.future import select
 from sqlalchemy.dialects.postgresql import Insert as _insert
 from sqlalchemy.exc import SQLAlchemyError
 
-from fidesapi import errors
+from fidesapi.utils import errors
 from fidesapi.database.session import async_session
-from fidesapi.sql_models import SqlAlchemyBase, sql_model_map
-from fideslang import model_map
-
-
-def get_resource_type(router: APIRouter) -> str:
-    "Extracts the name of the resource type from the prefix."
-    return router.prefix[1:]
+from fidesapi.sql_models import SqlAlchemyBase
 
 
 # CRUD Functions
@@ -186,67 +177,3 @@ async def delete_resource(sql_model: SqlAlchemyBase, fides_key: str) -> SqlAlche
                 raise error
 
         return sql_resource
-
-
-# CRUD Endpoints
-routers = []
-for resource_type, resource_model in model_map.items():
-    # Programmatically define routers for each resource type
-    router = APIRouter(
-        tags=[resource_model.__name__],
-        prefix=f"/{resource_type}",
-    )
-
-    @router.post(
-        "/", response_model=resource_model, status_code=status.HTTP_201_CREATED
-    )
-    async def create(
-        resource: resource_model,
-        resource_type: str = get_resource_type(router),
-    ) -> Dict:
-        """Create a resource."""
-        sql_model = sql_model_map[resource_type]
-        return await create_resource(sql_model, resource.dict())
-
-    @router.get("/", response_model=List[resource_model], name="List")
-    async def ls(  # pylint: disable=invalid-name
-        resource_type: str = get_resource_type(router),
-    ) -> List:
-        """Get a list of all of the resources of this type."""
-        sql_model = sql_model_map[resource_type]
-        return await list_resource(sql_model)
-
-    @router.get("/{fides_key}", response_model=resource_model)
-    async def get(
-        fides_key: str, resource_type: str = get_resource_type(router)
-    ) -> Dict:
-        """Get a resource by its fides_key."""
-        sql_model = sql_model_map[resource_type]
-        return await get_resource(sql_model, fides_key)
-
-    @router.put("/", response_model=resource_model)
-    async def update(
-        resource: resource_model,
-        resource_type: str = get_resource_type(router),
-    ) -> Dict:
-        """Update a resource by its fides_key."""
-        sql_model = sql_model_map[resource_type]
-        return await update_resource(sql_model, resource.dict())
-
-    @router.delete("/{fides_key}")
-    async def delete(
-        fides_key: str, resource_type: str = get_resource_type(router)
-    ) -> Dict:
-        """Delete a resource by its fides_key."""
-        sql_model = sql_model_map[resource_type]
-        deleted_resource = await delete_resource(sql_model, fides_key)
-        # Convert the resource to a dict explicitly for the response
-        deleted_resource_dict = (
-            model_map[resource_type].from_orm(deleted_resource).dict()
-        )
-        return {
-            "message": "resource deleted",
-            "resource": deleted_resource_dict,
-        }
-
-    routers += [router]
