+++ fidesctl/src/fidesapi/routes/visualize.py
@@ -7,7 +7,7 @@ from typing import Union
 from fastapi import APIRouter, HTTPException
 from fastapi.responses import HTMLResponse
 
-from fidesapi.crud import list_resource
+from fidesapi.routes.crud import list_resource
 from fidesapi.sql_models import sql_model_map
 from fidesctl.core import visualize
 from fideslang import model_map
