+++ fidesctl/src/fidesapi/view.py
@@ -3,7 +3,7 @@ Contains api endpoints for fides web pages
 """
 from fastapi import APIRouter
 from fastapi.responses import HTMLResponse
-from fidesapi.crud import list_resource
+from fidesapi.routes.crud import list_resource
 from fidesapi.sql_models import Evaluation
 
 router = APIRouter(
